CustomIdentityComponent/lambda/requirements.txt
pyjwt[crypto]
jwcrypto
requests
boto3
aws-xray-sdk
aws-lambda-powertools[aws-sdk]
